🏷️ #18 words 상태값 타입 및 타입 분리
